Create sub components to structure logic more
